lesson6/seminar/step2_function_calling/pyproject.toml
[tool.poetry]
name = "lesson6-step2"
version = "0.1.0"
description = "Function calling with Pose Visualization service"
authors = ["MLOps Course"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.12.0"
requests = "^2.31.0"
pydantic = "^2.5.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"
matplotlib = "^3.8.0"
numpy = "^1.24.0"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pytest = "^7.4.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
